Decompose large components for better maintainability
- Broke down ProjAndBlogClient.tsx (267 lines → 56 lines)
- Created focused BlogHeader component for title, date, GitHub link
- Added BlogContent component for article rendering
- Extracted ImageModal component for modal functionality
- Moved image modal logic to useImageModal hook
- Improved code reusability and maintainability
- Easier testing and debugging of individual components
